Added trajectory logging. Trajectories are now stored and processed in head angles instead of image space coordinates. Gesture recognition now compatible with head movement. Added lowpass filtering of trajectory points to reduce the effect of noise (and head movement) on gesture recognition. Gestures can now timeout, causing an interrupted or completed gesture, depending on state.

+++ include/ObjectTracking.hpp
@@ -61,7 +61,7 @@ class TrackedObject{
         RotatedRect useCamShift(const Mat probImage);
         double compare(boost::shared_ptr<TrackedObject> otherObject);
         void unOcclude();
-        void updateTrajectory(Point2f pt, float time);
+        void updateTrajectory(Point2f pt, long long time);
 };
 
 typedef map<int,boost::shared_ptr<TrackedObject> > objMap;
